Set up production Redux store
Set up a dev version and production version of the store, and
dynamically load the correct store in configureStore.js. Since it no
longer exports default, add a line in index.js to disable the linting
check for the default import.
